move the ColorType alias into the theme package
